Removed the clsx package. I changed parts of the temp drawer so that hooks were named more accurately. I also removed a the toggle function and had links directly push the drawer away. I also moved everything theme related to the app page. I made the accordion permanently expanded, but will have to change this once I add other segments with their own links such as the social section and gym section. - Denny Min
